Fixed resetTabList
Whenever the last row (variable b) is 3 an ArrayIndexOutOfBoundsException will happen. This is because the TabHolder holds a multidimensional array. Arrays will start from 0. This means the y variable in TabObject needs to be from 0-2.
Due to the point that the horzTabSize variable may be needed in other points the if statement is decrementing the variable and horzTabSize is not changed to 2 in general. (Since there are three rows)